backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Cors>=4.0
Flask-Sock>=0.7
psycopg[binary]>=3.1
python-dotenv>=1.0
pydantic>=2.8
orjson>=3.10
PyMuPDF>=1.24
Pillow>=10.2
numpy>=1.26
reportlab>=4.1
PyPDF2>=3.0
openai>=1.13
anthropic>=0.39
google-generativeai>=0.5
tenacity>=8.2
httpx>=0.27
redis>=5.0
python-json-logger>=2.0
structlog>=24.1
uvicorn>=0.30
